Use val_mgr to obtain Val objects
In addition to the benefits of using pre-allocated Val's, this resolves
issues with the Val(double) constructor being called for uint16
variables, which led to 19-digit decimal count values arriving in event
handlers.  One of those variables held the number of tags that was a
loop exit condition in a plugin main.zeek gquic_hello event handler,
which made it essentially infinite.